Add router and rename page
