+++ package.json
@@ -14,6 +14,7 @@
     "postcss": "^8.4.49",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^7.0.2",
     "tailwindcss": "^3.4.15"
   },
   "devDependencies": {
